Add sj_SavedLocation for push/pop of parse state
I'm not sure if you intended for the user to do this externally, but
with push/pop of the parse location, it makes it possible to scan an
object or array, then acquire memory for storing results, and then
reparse to process the data. e.g. something like:

static int count_entries(sj_Reader* reader, sj_Value obj) {
  sj_SavedLocation saved = sj_save_location(reader);
  sj_Value k, v;
  int count = 0;
  while (sj_iter_object(reader, obj, &k, &v)) {
    ++count;
  }
  sj_restore_location(reader, saved);
  return count;
}

int main() {
  // ...
  int count = count_entries(&reader, some_obj);
  MyParsedObj* parsed = malloc(sizeof(MyParsedObj) * count);
  while (sj_iter_object(reader, some_obj, &k, &v)) {
    // ... fill out parsed
  }
}